cifar10-wip

+++ cifar-10/src/input_pipeline.py
@@ -56,11 +56,14 @@ class DataPipeline(object):
                             capacity=self._batch_size*3,
                             min_after_dequeue=self._batch_size*2
                         )
-        
+
         ret_x = tf.transpose(
                     tf.reshape(x, [self._batch_size, 3, self._w, self._h]),
-                    (0, 2, 3, 1) 
+                    (0, 2, 3, 1)
         )
-        ret_y = tf.one_hot(y, self._num_labels, 1, 0)
+        ret_y = tf.cast(
+            tf.one_hot(y, self._num_labels, 1, 0),
+            tf.float32
+        )
 
         return ret_x, ret_y
